Reflecting our new-newest understanding
This is what the system will look like if we're using Prefect Server/Cloud as
the broker, and using Prefect's results system for task argument and result
storage.  In this scenario, the `api` and `tasks` deployments don't need to know
about each other or communicate directly, and the `TaskServer` does not run any
kind of external API.  Both `api` and `tasks` need access to the same storage,
represented here as sharing a Docker volume (and thus LocalFileSystem storage).