Beefing up the test suite with more realistic tests
My goal here is to write the kinds of tests an outside dev would want to write,
just like we would want to do with our test suite: making an isolated
environment, queueing up tasks, then running a single loop of the task server.
This models what we current do in Nebula with `arq`.