Adding the new `chaos-duck` example app, for testing task server reliability
This example application starts a system with simple tasks to run, then randomly
fails in various fun ways.  The goal is for the `__main__.py` script to always
run to completion and show that all tasks eventually end up `COMPLETED`.  We
aren't quite there 100% of the time, but it's heading that way.